Rename package to flask-restalchemy
The new name makes more explicit that the package is intended to work
with SQLAlchemy models.
