+++ flask_restalchemy/resources/docgeneration.py
